Cambios de confirmacion

+++ resources/views/admin/menu-conocenos/perfil-egreso/editar-perfil-ingreso.blade.php
@@ -17,7 +17,7 @@
 		<div class="formulario">
 			
 			<label for="">Descripción Perfil Egreso</label>
-			<a href="{{route('crear-perfil-egreso')}}"><button class="btn"><span>Agregar</span></button></a>
+			<a onclick="return confirm('¿Desea agregar un nuevo elemento?')" href="{{route('crear-perfil-egreso')}}"><button class="btn"><span>Agregar</span></button></a>
 
 			
 		</div>
@@ -44,8 +44,8 @@
 							<tr>
 								<td>{{$perfilegres->vineta}}</td>
 								<td>{{$perfilegres->elemento}}</td>
-								<td><a href="eliminar-elemento-eg/{{$perfilegres->slug}}"><button class="btn-tabla"><i class="far fa-trash-alt"></i></button></a></td>
-								<td><a href="/NuevoElementoEg/{{$perfilegres->slug}}/edit"><button class="btn-tabla"><i class="fas fa-pencil-alt"></i></button></a></td>	
+								<td><a onclick="return confirm('¿Desea eliminar este elemento?')" href="eliminar-elemento-eg/{{$perfilegres->slug}}"><button class="btn-tabla"><i class="far fa-trash-alt"></i></button></a></td>
+								<td><a onclick="return confirm('¿Desea editar este elemento?')" href="/NuevoElementoEg/{{$perfilegres->slug}}/edit"><button class="btn-tabla"><i class="fas fa-pencil-alt"></i></button></a></td>	
 							</tr>
 
 							@endforeach
